+++ fixed-vector/Data/Vector/Fixed/Boxed.hs
@@ -1,14 +1,5 @@
-{-# LANGUAGE DataKinds             #-}
-{-# LANGUAGE DeriveDataTypeable    #-}
-{-# LANGUAGE FlexibleContexts      #-}
-{-# LANGUAGE FlexibleInstances     #-}
-{-# LANGUAGE MagicHash             #-}
-{-# LANGUAGE MultiParamTypeClasses #-}
-{-# LANGUAGE ScopedTypeVariables   #-}
-{-# LANGUAGE StandaloneDeriving    #-}
-{-# LANGUAGE TypeApplications      #-}
-{-# LANGUAGE TypeFamilies          #-}
-{-# LANGUAGE UndecidableInstances  #-}
+{-# LANGUAGE MagicHash            #-}
+{-# LANGUAGE UndecidableInstances #-}
 -- |
 -- Vector which could hold any value.
 module Data.Vector.Fixed.Boxed (
@@ -35,14 +26,12 @@ import Foreign.Storable (Storable(..))
 import GHC.TypeLits
 import GHC.Exts (proxy#)
 import Prelude ( Show(..),Eq(..),Ord(..),Functor(..),Monad(..)
-               , ($),($!),error,seq)
+               , ($),($!),error)
 
 import Data.Vector.Fixed hiding (index)
 import Data.Vector.Fixed.Mutable (Mutable, MVector(..), IVector(..), DimM, constructVec, inspectVec, index)
 import qualified Data.Vector.Fixed.Cont     as C
 import           Data.Vector.Fixed.Cont     (Peano,ArityPeano(..))
-import qualified Data.Vector.Fixed.Internal as I
-
 
 
 ----------------------------------------------------------------
@@ -55,53 +44,38 @@ newtype Vec (n :: Nat) a = Vec (SmallArray a)
 -- | Mutable unboxed vector with fixed length
 newtype MVec (n :: Nat) s a = MVec (SmallMutableArray s a)
 
-deriving instance Typeable Vec
-deriving instance Typeable MVec
-
 type Vec1 = Vec 1
 type Vec2 = Vec 2
 type Vec3 = Vec 3
 type Vec4 = Vec 4
 type Vec5 = Vec 5
 
-
-instance (Typeable n, Arity n, Data a) => Data (Vec n a) where
-  gfoldl       = C.gfoldl
-  gunfold      = C.gunfold
-  toConstr   _ = con_Vec
-  dataTypeOf _ = ty_Vec
-
-ty_Vec :: DataType
-ty_Vec  = mkDataType "Data.Vector.Fixed.Boxed.Vec" [con_Vec]
-
-con_Vec :: Constr
-con_Vec = mkConstr ty_Vec "Vec" [] Prefix
-
-instance (Storable a, Arity n) => Storable (Vec n a) where
-  alignment = defaultAlignemnt
-  sizeOf    = defaultSizeOf
-  peek      = defaultPeek
-  poke      = defaultPoke
-  {-# INLINE alignment #-}
-  {-# INLINE sizeOf    #-}
-  {-# INLINE peek      #-}
-  {-# INLINE poke      #-}
-
-
+type instance Mutable (Vec  n) = MVec n
+type instance Dim     (Vec  n) = Peano n
+type instance DimM    (MVec n) = Peano n
 
 
 ----------------------------------------------------------------
 -- Instances
 ----------------------------------------------------------------
 
-instance (Arity n, Show a) => Show (Vec n a) where
-  showsPrec = I.showsPrec
+deriving via ViaFixed (Vec n) instance Arity n => Functor    (Vec n)
+deriving via ViaFixed (Vec n) instance Arity n => Applicative (Vec n)
+deriving via ViaFixed (Vec n) instance Arity n => F.Foldable  (Vec n)
 
-instance (Arity n, NFData a) => NFData (Vec n a) where
-  rnf = foldl (\r a -> r `seq` rnf a) ()
-  {-# INLINE rnf #-}
+instance Arity n => T.Traversable (Vec n) where
+  sequenceA = sequenceA
+  traverse  = traverse
+  {-# INLINE sequenceA #-}
+  {-# INLINE traverse #-}
 
-type instance Mutable (Vec n) = MVec n
+deriving via ViaFixed (Vec n) a instance (Arity n, Show      a) => Show      (Vec n a)
+deriving via ViaFixed (Vec n) a instance (Arity n, Eq        a) => Eq        (Vec n a)
+deriving via ViaFixed (Vec n) a instance (Arity n, Ord       a) => Ord       (Vec n a)
+deriving via ViaFixed (Vec n) a instance (Arity n, NFData    a) => NFData    (Vec n a)
+deriving via ViaFixed (Vec n) a instance (Arity n, Semigroup a) => Semigroup (Vec n a)
+deriving via ViaFixed (Vec n) a instance (Arity n, Monoid    a) => Monoid    (Vec n a)
+deriving via ViaFixed (Vec n) a instance (Arity n, Storable  a) => Storable  (Vec n a)
 
 instance (Arity n) => MVector (MVec n) a where
   new = do
@@ -125,11 +99,6 @@ instance (Arity n) => IVector (Vec n) a where
   {-# INLINE unsafeThaw   #-}
   {-# INLINE unsafeIndex  #-}
 
-
-
-type instance Dim  (Vec  n) = Peano n
-type instance DimM (MVec n) = Peano n
-
 instance (Arity n) => Vector (Vec n) a where
   construct  = constructVec
   inspect    = inspectVec
@@ -138,42 +107,17 @@ instance (Arity n) => Vector (Vec n) a where
   {-# INLINE inspect    #-}
   {-# INLINE basicIndex #-}
 
-instance (Arity n, Eq a) => Eq (Vec n a) where
-  (==) = eq
-  {-# INLINE (==) #-}
-instance (Arity n, Ord a) => Ord (Vec n a) where
-  compare = ord
-  {-# INLINE compare #-}
+instance (Typeable n, Arity n, Data a) => Data (Vec n a) where
+  gfoldl       = C.gfoldl
+  gunfold      = C.gunfold
+  toConstr   _ = con_Vec
+  dataTypeOf _ = ty_Vec
 
-instance (Arity n, Monoid a) => Monoid (Vec n a) where
-  mempty  = replicate mempty
-  mappend = (<>)
-  {-# INLINE mempty  #-}
-  {-# INLINE mappend #-}
+ty_Vec :: DataType
+ty_Vec  = mkDataType "Data.Vector.Fixed.Boxed.Vec" [con_Vec]
 
-instance (Arity n, Semigroup a) => Semigroup (Vec n a) where
-  (<>) = zipWith (<>)
-  {-# INLINE (<>) #-}
-
-instance Arity n => Functor (Vec n) where
-  {-# INLINE fmap #-}
-  fmap = map
-
-instance Arity n => Applicative (Vec n) where
-  pure  = replicate
-  (<*>) = zipWith ($)
-  {-# INLINE pure  #-}
-  {-# INLINE (<*>) #-}
-
-instance Arity n => F.Foldable (Vec n) where
-  foldr = foldr
-  {-# INLINE foldr #-}
-
-instance Arity n => T.Traversable (Vec n) where
-  sequenceA = sequenceA
-  traverse  = traverse
-  {-# INLINE sequenceA #-}
-  {-# INLINE traverse #-}
+con_Vec :: Constr
+con_Vec = mkConstr ty_Vec "Vec" [] Prefix
 
 uninitialised :: a
 uninitialised = error "Data.Vector.Fixed.Boxed: uninitialised element"